backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
pydantic==2.10.5
pydantic-settings==2.7.1
python-multipart==0.0.20
openpyxl==3.1.5
aiogram==3.16.0
python-dotenv==1.0.1
sqladmin==0.20.1
python-slugify==8.0.4
greenlet==3.1.1
httpx==0.28.1
pandas==2.2.3
xlrd==2.0.1
